PSCF2-81 - Renamed it.gov.pagopa.swclient.mil to it.pagopa.swclient.mil
