+++ src/main/java/it/pagopa/swclient/mil/ErrorCode.java
@@ -3,7 +3,7 @@
  *
  * 26 nov 2022
  */
-package it.gov.pagopa.swclient.mil;
+package it.pagopa.swclient.mil;
 
 /**
  *
